amend

+++ torchrl/envs/libs/envpool.py
@@ -150,7 +150,12 @@ class MultiThreadedEnvWrapper(_EnvWrapper):
     def _step(self, tensordict: TensorDictBase) -> TensorDictBase:
         action = tensordict.get(self.action_key)
         if self.xla:
-            self._env_handle, step_output = self._step_jax(self._env_handle, jax.dlpack.from_dlpack(torch.utils.dlpack.to_dlpack(action.to(torch.int32))))
+            self._env_handle, step_output = self._step_jax(
+                self._env_handle,
+                jax.dlpack.from_dlpack(
+                    torch.utils.dlpack.to_dlpack(
+                        action
+                    )))
         else:
             # Action needs to be moved to CPU and converted to numpy before being passed to envpool
             action = action.to(torch.device("cpu"))
